+++ packages/expo-notifications/android/src/main/java/expo/modules/notifications/notifications/emitting/NotificationsEmitter.kt
@@ -41,6 +41,11 @@ open class NotificationsEmitter : Module(), NotificationListener {
     AsyncFunction<Bundle?>("getLastNotificationResponseAsync") {
       lastNotificationResponseBundle
     }
+
+    AsyncFunction("clearLastNotificationResponseAsync") {
+      lastNotificationResponseBundle = null
+      null
+    }
   }
 
   /**